Check if vouchers are available on cart_footer.tpl
With the settings you can say, that no additional options should be shown on checkout. This means that for example no voucher shoul be possible. When you change this setting to "no" vouchers will be shown on cart. But I think that is not correct.
To change this behaviour I added the "if" to cart_footer.tpl